Restructure tests to mirror source code layout
Move flat test files into a directory structure that mirrors core/:
- tests/core/adapters/amp/ for AMP adapter tests
- tests/core/adapters/ for base adapter tests
- tests/core/game/ for state manager tests
- tests/core/network/ for network utility tests
- tests/integration/ for tests requiring external services
- tests/tui/ for TUI tests

Rename *_test.py to test_*.py (pytest convention), fix stale
obs_controller imports, remove obsolete dota2 rcon test, and
add conftest.py with shared fixtures.
