+++ tests/core/adapters/amp/test_message_processor.py
