+++ tests/core/adapters/amp/test_status_parser.py
