+++ tests/core/adapters/test_message_processor.py
