+++ tests/core/adapters/test_message_type.py
