+++ tests/core/game/test_state_manager.py
