+++ tests/integration/test_amp_console_stream.py
